10-10d | Reorganize utils structure
