+++ src/applications/ivc-champva/10-10d-extended/chapters/applicantInformation/remarriageProof.js
@@ -3,7 +3,7 @@ import {
   descriptionUI,
 } from 'platform/forms-system/src/js/web-component-patterns';
 import RemarriageProofDescription from '../../components/FormDescriptions/RemarriageProofDescription';
-import { ATTACHMENT_IDS } from '../../constants';
+import { ATTACHMENT_IDS } from '../../utils/constants';
 import { attachmentSchema, attachmentUI } from '../../definitions';
 import content from '../../locales/en/content.json';
