+++ src/applications/ivc-champva/10-10d-extended/components/FormDescriptions/HealthInsuranceSummaryCard.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import { format, parseISO } from 'date-fns';
 import { SCHEMA_LABELS } from '../../chapters/healthInsuranceInformation/planTypes';
-import { generateParticipantNames } from '../../helpers';
+import { generateParticipantNames } from '../../utils/helpers';
 import content from '../../locales/en/content.json';
 
 const TEXT_PRESENT = content['health-insurance--summary-date-present'];
