+++ src/applications/ivc-champva/10-10d-extended/components/FormFooter/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { CONTACTS } from '../../imports';
+import { CONTACTS } from '../../utils/imports';
 
 const GetFormHelp = (
   <>
