+++ src/applications/ivc-champva/10-10d-extended/config/submitTransformer.js
@@ -5,7 +5,7 @@ import {
   getObjectsWithAttachmentId,
   toHash,
 } from '../../shared/utilities';
-import { concatStreets, getAgeInYears } from '../helpers/utilities';
+import { concatStreets, getAgeInYears } from '../utils/helpers';
 
 /**
  * Formats a date string from YYYY-MM-DD to MM-DD-YYYY
