+++ src/applications/ivc-champva/10-10d-extended/tests/unit/utils/helpers/formatting.unit.spec.jsx
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { replaceStrValues } from '../../../helpers/formatting';
+import { replaceStrValues } from '../../../../utils/helpers';
 
 describe('10-10d `replaceStrValues` util', () => {
   it('should return empty string when source is falsy', () => {
