+++ src/applications/ivc-champva/10-10d-extended/tests/unit/utils/helpers/helpers.unit.spec.jsx
@@ -4,7 +4,7 @@ import {
   getAgeInYears,
   page15aDepends,
   populateFirstApplicant,
-} from '../../../helpers/utilities';
+} from '../../../../utils/helpers';
 
 describe('page15a depends function', () => {
   const isApp = {
